fix(supplement): use Integer id to match Supabase serial (not bigint)
Made-with: Cursor

+++ src/main/java/com/memi/lifeos/supplement/entity/Supplement.java
@@ -25,7 +25,8 @@ public class Supplement {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	/** Maps to PostgreSQL/Supabase {@code serial} (INTEGER), not BIGSERIAL. */
+	private Integer id;
 
 	@Column(nullable = false)
 	private String name;
